refactor: split rates supply snapshot builder
